fix tests 	1. split test code into files according to routes 	2. route handler now use an interface, instead of the struct contains db instacne, this modification is made for the convenience of test
