+++ routes/article.go
@@ -13,7 +13,7 @@ type articleHandler struct{}
 func (r *articleHandler) ArticleGetHandler(c *gin.Context) {
 
 	input := models.Article{ID: c.Param("id")}
-	article, err := models.DB.Get(input)
+	article, err := models.DS.Get(input)
 
 	if err != nil {
 		switch err.Error() {
@@ -51,7 +51,7 @@ func (r *articleHandler) ArticlePostHandler(c *gin.Context) {
 	if article.Active != 1 {
 		article.Active = 1
 	}
-	result, err := models.DB.Create(article)
+	result, err := models.DS.Create(article)
 	if err != nil {
 		switch err.Error() {
 		case "Duplicate entry":
@@ -82,7 +82,7 @@ func (r *articleHandler) ArticlePutHandler(c *gin.Context) {
 		article.UpdatedAt.Time = time.Now()
 		article.UpdatedAt.Valid = true
 	}
-	result, err := models.DB.Update(article)
+	result, err := models.DS.Update(article)
 	if err != nil {
 		switch err.Error() {
 		case "Article Not Found":
@@ -100,7 +100,7 @@ func (r *articleHandler) ArticleDeleteHandler(c *gin.Context) {
 
 	input := models.Article{ID: c.Param("id")}
 	// var req models.Databox = &models.Member{ID: userID}
-	article, err := models.DB.Delete(input)
+	article, err := models.DS.Delete(input)
 
 	// member, err := req.Delete()
 	if err != nil {
